requirements.txt
requests
beautifulsoup4
apscheduler
python-telegram-bot==21.*
playwright
